34580 & 34581
Evasive state updated to not circle around the player anymore, teh AI
now make distance between itself and the player and tries to maintain
this distance for the remained of the states predetermined duration. A
QuickBlock state has also been added, this allows the AI to block a
players attack while they're attacking, the trigger for this is
currently eveyrtime the player attacks however a better trigger for this
state will need to be found.

The IAIAttribute interface was also added so as enable the creation of a
list that can be passed through the state machine to avoid unneccessary
'GetComponent' calls and thus making state transitions less expensive.
The interface is currently empty and however it is best if not every
MonoBehaviour gets added to this list so using an empty interface is the
best alternative.