Fixed states still using GetComponent
Several states were not updated to use the attributes list, also most
states were using GetComponent to reach the Animator so the animator is
now something that is passed through the constructor of every state.
This allows for redueced time switching between states as the states do
not need to get retrieve the animator.