34583
Added checks for enemy health when deciding which patterns to use and
added probability so that the patterns used aren't always the same, this
required some playing around with and testing to make sure it was
switching patterns correctly, the actual patterns still need work done
to decide which ones are most appropriate where and when. The patterns
may ultimatly have to be completly redone when the new animations and
models are implemented.